Telefonbuch Challange
A friend gave me a challange to further enhance my coding skills here is the result. The Quest was to creat a Simple Phone book where you can Store new Entrys or Delet old ones or have a insight into the whole list. The list than needs to be saved into a external file that saves the changes but without using a libary